+++ src/xrGame/MainMenu.h
@@ -149,7 +149,7 @@ public:
     void IR_OnMousePress(int btn) override;
     void IR_OnMouseRelease(int btn) override;
     void IR_OnMouseHold(int btn) override;
-    void IR_OnMouseWheel(int x, int y) override;
+    void IR_OnMouseWheel(float x, float y) override;
     void IR_OnMouseMove(int x, int y) override;
 
     void IR_OnKeyboardPress(int dik) override;
