+++ src/xrGame/UIDialogHolder.h
@@ -65,7 +65,7 @@ public:
     virtual bool IgnorePause() { return false; }
 
     virtual bool IR_UIOnMouseMove(int dx, int dy);
-    virtual bool IR_UIOnMouseWheel(int x, int y);
+    virtual bool IR_UIOnMouseWheel(float x, float y);
 
     virtual bool IR_UIOnKeyboardPress(int dik);
     virtual bool IR_UIOnKeyboardRelease(int dik);
